:fire: Remove unused assets and files
