--- Assets/Materials/RealTimeMapMat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a48ffceec711b2d44aee751b43345327
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
