+++ Assets/Materials/RealTimePaintMat.mat.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bfed933cd36d1b04a9aac185fbc5b0db
+guid: 40eb76054e605ec45b1a61c1a2c29232
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 2100000
