--- Assets/Materials/RealTimePaintMat_.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 05028551ca67fd847921d0a31b1d46a2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
