--- Assets/Materials/RealTimePaintMat__.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 40eb76054e605ec45b1a61c1a2c29232
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
