--- Assets/Materials/SampleMat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7d04d71e08ff9c84aafacecd388c5244
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
