--- Assets/Prefabs/Quad.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2001af3bd65b4704d96df352459c5d41
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
