--- Assets/Scripts/RealTimeMap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f1e02dc2744b65d4e9883f648ca13f53
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
