+++ Assets/Scripts/RealTimePaint.cs
@@ -1,79 +1,125 @@
-﻿using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
+using System.Collections;
 
 public class RealTimePaint : MonoBehaviour
 {
     [SerializeField]
-    private Material paintMaterial = null;
-    private RenderTexture paintTexture = null;
+    private Renderer myRenderer = null;
     private int paintWorldPositionId = 0;
-    private int viewProjectionInverseId = 0;
+    private int prevTexId = 0;
+    private int currentTexId = 0;
+    /// <summary>
+    /// 前フレームのテクスチャを参照するためのテクスチャ．
+    /// paintBuffer2と書き込み，読み込み役を毎フレーム交代する．
+    /// </summary>
+    private RenderTexture paintBuffer1 = null;
+    /// <summary>
+    /// 前フレームのテクスチャを参照するためのテクスチャ．
+    /// paintBuffer1と書き込み，読み込み役を毎フレーム交代する．
+    /// </summary>
+    private RenderTexture paintBuffer2 = null;
+    /// <summary>
+    /// 偶数フレームで描画されたかどうか
+    /// </summary>
+    private bool isEvenFrameRendered = false;
+    private Vector4 worldPosition = new Vector4(0, 0, 0, -1);
+    // [SerializeField]
+    // private Renderer buffer1 = null;
+    // [SerializeField]
+    // private Renderer buffer2 = null;
 
     void Awake()
     {
-        Renderer renderer = GetComponent<Renderer>();
-        Texture mainTexture = renderer.material.mainTexture;
+        Texture mainTexture = myRenderer.material.mainTexture;
 
-        // RenderTextureの生成
-        paintTexture = new RenderTexture(mainTexture.width, mainTexture.height, 0, RenderTextureFormat.ARGBFloat, RenderTextureReadWrite.Default);
-        paintTexture.enableRandomWrite = true;
+        // 前フレームのテクスチャ参照用
+        paintBuffer1 = new RenderTexture(1024, 1024, 0, RenderTextureFormat.ARGB32, RenderTextureReadWrite.Default);
+        paintBuffer2 = new RenderTexture(1024, 1024, 0, RenderTextureFormat.ARGB32, RenderTextureReadWrite.Default);
+        paintBuffer1.enableRandomWrite = true;
+        paintBuffer2.enableRandomWrite = true;
 
-        // 元々のテクスチャを生成したRenderTextureにコピー
-        Graphics.Blit(mainTexture, paintTexture);
-
-        // paintMaterialのテクスチャに設定
-        int mainTextureId = Shader.PropertyToID("_MainTex");
-        paintMaterial.SetTexture(mainTextureId, paintTexture);
-
-        // MainTextureを生成したRenderTextureに変更する
-        renderer.material.mainTexture = paintTexture;
+        // メインテクスチャで初期化
+        Graphics.Blit(mainTexture, paintBuffer1);
+        Graphics.Blit(mainTexture, paintBuffer2);
 
         // 各シェーダープロパティIDを取得
         paintWorldPositionId = Shader.PropertyToID("_PaintWorldPosition");
-        viewProjectionInverseId = Shader.PropertyToID("_ViewProjectionInverse");
+        prevTexId = Shader.PropertyToID("_PrevTex");
+        currentTexId = Shader.PropertyToID("_CurrentTex");
 
-        // シェーダーでdepthを取り出すための設定
-        Camera.main.depthTextureMode = DepthTextureMode.Depth;
+        StartCoroutine(UpdateTextureAction());
+
+        // // for debug
+        // buffer1.material.mainTexture = paintBuffer1;
+        // buffer2.material.mainTexture = paintBuffer2;
     }
 
     void Update()
     {
-        Paint();
+        // Updateメソッドはレンダリング前に実行されるため，必要なデータはここでセットする．
+        UpdatePaint();
     }
 
-    // void OnRenderImage(RenderTexture src, RenderTexture dest)
-    // {
-    //     paintMaterial.SetVector(paintWorldPositionId, new Vector4(0, 0, 0, 0));
-
-    //     // from https://gamedev.stackexchange.com/questions/131978/shader-reconstructing-position-from-depth-in-vr-through-projection-matrix/140924#140924
-    //     Matrix4x4 projectionMatrix = GL.GetGPUProjectionMatrix(Camera.main.projectionMatrix, false);
-    //     projectionMatrix[2, 3] = projectionMatrix[3, 2] = 0.0f;
-    //     projectionMatrix[3, 3] = 1.0f;
-
-    //     Matrix4x4 clipToWorld = Matrix4x4.Inverse(projectionMatrix * Camera.main.worldToCameraMatrix) * Matrix4x4.TRS(new Vector3(0, 0, -projectionMatrix[2, 2]), Quaternion.identity, Vector3.one);
-    //     paintMaterial.SetMatrix(viewProjectionInverseId, clipToWorld);
-
-    //     // ペイント処理
-    //     Graphics.Blit(src, dest, paintMaterial);
-    // }
-
-    void Paint()
+    public void Paint(Vector3 worldPosition)
     {
-        paintMaterial.SetVector(paintWorldPositionId, new Vector4(0, 0, 0, 0));
+        this.worldPosition = new Vector4(worldPosition.x, worldPosition.y, worldPosition.z, 1);
+    }
 
-        // from https://gamedev.stackexchange.com/questions/131978/shader-reconstructing-position-from-depth-in-vr-through-projection-matrix/140924#140924
-        Matrix4x4 projectionMatrix = GL.GetGPUProjectionMatrix(Camera.main.projectionMatrix, false);
-        projectionMatrix[2, 3] = projectionMatrix[3, 2] = 0.0f;
-        projectionMatrix[3, 3] = 1.0f;
+    /// <summary>
+    /// リアルタイムペイントを行う．
+    /// </summary>
+    void UpdatePaint()
+    {
+        // シェーダーにペイントするワールド座標を設定する
+        myRenderer.material.SetVector(paintWorldPositionId, worldPosition);
 
-        Matrix4x4 clipToWorld = Matrix4x4.Inverse(projectionMatrix * Camera.main.worldToCameraMatrix) * Matrix4x4.TRS(new Vector3(0, 0, -projectionMatrix[2, 2]), Quaternion.identity, Vector3.one);
-        paintMaterial.SetMatrix(viewProjectionInverseId, clipToWorld);
+        if (Time.frameCount % 2 == 0)
+        {
+            // 偶数フレームはpaintBuffer1を保存用，paintBuffer2を書き込み用にする
+            myRenderer.material.SetTexture(prevTexId, paintBuffer1);
+            myRenderer.material.SetTexture(currentTexId, paintBuffer2);
+            isEvenFrameRendered = true;
 
-        // ペイント処理
-        RenderTexture tmp = RenderTexture.GetTemporary(paintTexture.width, paintTexture.height, 0, RenderTextureFormat.ARGBFloat, RenderTextureReadWrite.Default);  // 一時的なテクスチャを作成する
-        Graphics.Blit(paintTexture, tmp, paintMaterial);  // paintMaterialを基にtmpに描画する
-        Graphics.Blit(tmp, paintTexture); // その後，tmpに描画された結果をpaintTextureにコピー
-        RenderTexture.ReleaseTemporary(tmp);  // tmpを解放
+            // RWTexture2Dに書き込みを行うために必要
+            Graphics.ClearRandomWriteTargets();
+            Graphics.SetRandomWriteTarget(1, paintBuffer2);
+        }
+        else
+        {
+            // 奇数フレームはpaintBuffer2を保存用，paintBuffer1を書き込み用にする
+            myRenderer.material.SetTexture(prevTexId, paintBuffer2);
+            myRenderer.material.SetTexture(currentTexId, paintBuffer1);
+            isEvenFrameRendered = false;
+
+            // RWTexture2Dに書き込みを行うために必要
+            Graphics.ClearRandomWriteTargets();
+            Graphics.SetRandomWriteTarget(1, paintBuffer1);
+        }
+    }
+
+    /// <summary>
+    /// テクスチャを更新する処理
+    /// </summary>
+    /// <returns></returns>
+    IEnumerator UpdateTextureAction()
+    {
+        while (true)
+        {
+            // スクリーン上のレンダリングが完了するまで待つ
+            yield return new WaitForEndOfFrame();
+
+            if (isEvenFrameRendered)
+            {
+                // paintBuffer2に入っている新しい値でpaintBuffer1を更新する．
+                Graphics.Blit(paintBuffer2, paintBuffer1);
+            }
+            else
+            {
+                Graphics.Blit(paintBuffer1, paintBuffer2);
+            }
+
+            // 次フレームではペイントしないようにするための処理
+            worldPosition.w = -1;
+        }
     }
 }
