+++ Assets/Scripts/RealTimePaint.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b232a73bff01264dbe9dabc6a14d63a
+guid: 8045427a9038c3640b81db9d037db07c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
