--- Assets/Scripts/RealTimePaint_.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8045427a9038c3640b81db9d037db07c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
