--- Assets/Shaders/RealTimeMap.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 989abd79bb418594fa8cb3e19a3698df
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
