+++ Assets/Shaders/RealTimePaint.shader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fa936de624d102b4c9d4a3942bec7367
+guid: 0a643100e31c5404eb54a57a4f92679f
 ShaderImporter:
   externalObjects: {}
   defaultTextures: []
