--- Assets/Shaders/RealTimePaint_.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 79ba7f0268901804d9f77b3024cbe64c
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
