--- Assets/Shaders/RealTimePaint__.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 0a643100e31c5404eb54a57a4f92679f
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
